draw2 deleted, command.txt added, draw.cpp modified
The executable draw2 is no longer needed.
command.txt is the command used in order to build the program with
OpenCV.

draw.cpp print the matrix as an image with openCV and some point
operartors are commented.

The point operators are:
Negative (changes the values for the oposite ones)
Binary (less than 128 are changed by 0 and greater than 128 are
changed by 255)
Clarify (add 100 to every pixel)
Darkify (substract 100 to every pixel)