Functions improved in matrix_game.cpp
The way temp object is "repainted" was improved. It is no longer
creating new objects so is not wasting memory anymore.

The functions were improved so canvas is no longer needed as a
parameter in drawing functions. It allows you to re-use those
functions in other sections of the code.